Half-working attempt at Git ignoring
This does not work properly yet!

I’m having trouble getting the libgit2 ‘status_should_ignore’ function to work properly. I think there’s a mismatch between the paths that we give it and the paths that it expects, and for some reason it’s returning ‘true’ for paths that don’t exist in the repo.

To get this far, I had to add the working directory and the repository as fields inside the ‘Git’ struct. This is because we need to keep the repository around so we can query it for every file. Previously, the list of statuses was calculated up-front and held in memory from then on, which meant we didn’t need to deal with accessing the repository inside a ‘Mutex’ or any of that stuff.